feat: AIP-193
This adds a generic AIP for errors. There is probably a decent bit
for us to discuss here.

Some high-level notes:

- I decided to represent expected JSON interfaces using TypeScript
  (rather than JSONSchema), which I perceive to be much better
  for human readability.
- We should discuss/debate the proposed Error interface. It is
  mostly similar to what Google uses but with two fields "promoted"
  (we have a huge _mea culpa_ here).
- I did not discuss any common _headers_ related to error handling
  (e.g. `Retry-After`). I personally think that `Retry-After` gets
  covered in AIP-194, and I could not think of any others that
  warranted inclusion here.

I expect this is an area where everyone will need to make changes,
but I also notice that entire sections can probably be adopted
by everyone (e.g. "Messages"), so I think this should work
reasonably well.

Looking forward to the discussion on this.