feat: AIP-165 – Criteria-based delete
This adds a refactored AIP-165.

I really like some of the precedents set in #2; for example,
mirroring the structure there required me to add a full section
on errors that I consider to be quite useful. Similarly, separating
the behavior from the IDL still works well.

There is an open question here about long-running operations, which
I do not know how to properly handle in OAS and need to discuss
with @mkistler.