Clean up AIP-136: Custom methods. (#103)
* Clean up AIP-136: Custom methods.

This is in preparation for the barn raising event, and makes the AEP more consistent with more recently-adopted AEPs.  Common guidance is also factored out of the IDL-specific tabs.

I added one new bullet point of guidance while I was at it.  I doubt this will be controversial, but calling it out for completeness.

I actually think most of the changes from the Google AIP are pretty good in this case, so I have generally not modified them.

* Update creation date

* Update aep/general/0136/aep.md.j2

Co-authored-by: Yusuke Tsutsumi <yusuke@tsutsumi.io>

* Update aep/general/0136/aep.md.j2

Co-authored-by: Yusuke Tsutsumi <yusuke@tsutsumi.io>

* Address PR comments

---------

Co-authored-by: Yusuke Tsutsumi <yusuke@tsutsumi.io>