feat(errors): Add AEP-193 Errors (#158)
This PR adds AEP-193 that describes the standard structure and contents of an error response. It is significantly revised from the Google AIP-193 to adopt the HTTP standard "Problem Details" structure as defined in RFC 9547. However, I think the structure is largely compatible with the Google error response structure -- mostly differences in property names -- so I have left the proto-specific guidance from the AIP-193 in place.

I also dropped some of the details from AIP-193. In this particular AEP, I think "less is more". Let's not get bogged down in details. I think we debated details on this AEP for over a year and never got it adopted. Let's get the basics right here and leave details for later (if ever).