fix(133): add oas examples
Adding oas examples for create.

This also tests out the pattern outlined in #230 - 
adding a generated oas.yaml and use selective elements from it
as an example, which looks to work pretty well.

There is a blocker on https://github.com/aep-dev/site-generator-beta/issues/43
to allow for support from a json-path like structure to target sub-elements, but
this will at least allow for selective guidance to be easily added.